refactor(nms): Consolidate FEGSubscriberContext
